fix: id of content to be unique
`content` is already used as an id in the server template:
https://github.com/nextcloud/server/blob/c10317f7f9f53a4de464915f754896eed7f1ee6c/core/templates/layout.user.php#L125

Make sure to have a unique id and to replace the correct dom element when mounting the vue app.

The app uses NcAppContent which is rendered as a `<main>` tag.
Mount the app on `main#content` from the server template
so there is only one `<main>` on the page.

Signed-off-by: Max <max@nextcloud.com>